migrated to postgres

+++ izbabot/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from sqlalchemy.orm import declarative_base
 
 from utils import get_beer_word
@@ -14,7 +14,7 @@ class Member(Base):
         self.member_id = member_id
         self.name = name
 
-    member_id = Column(Integer, primary_key=True)
+    member_id = Column(BigInteger, primary_key=True)
     name = Column(String, nullable=False)
 
     def __repr__(self):
@@ -29,8 +29,8 @@ class OwnedBeer(Base):
         self.beer_to_id = int(beer_to_id)
         self.count = count
 
-    beer_from_id = Column(Integer, ForeignKey('members.member_id'), primary_key=True)
-    beer_to_id = Column(Integer, ForeignKey('members.member_id'), primary_key=True)
+    beer_from_id = Column(BigInteger, ForeignKey('members.member_id'), primary_key=True)
+    beer_to_id = Column(BigInteger, ForeignKey('members.member_id'), primary_key=True)
     count = Column(Integer, nullable=False)
 
     def __str__(self):
